requirements.txt
Django==1.8.3
django-allauth==0.22.0
oauthlib==1.0.0
pep8==1.6.2
python-openid==2.2.5
requests==2.7.0
requests-oauthlib==0.5.0
